+++ src/app/terra-components.module.ts
@@ -145,8 +145,6 @@ import { RouterModule } from '@angular/router';
 import { TerraTimePickerComponentExample } from './components/forms/input/time-picker/example/terra-time-picker.component.example';
 import { TerraTwoColumnsContainerComponent } from './components/layouts/column-container/terra-two-columns-container.component';
 import { CategoryTreeConfig } from './components/category-picker/config/category-tree.config';
-import { TerraBreadcrumbsComponent } from './components/breadcrumbs/terra-breadcrumbs.component';
-import { TerraBreadcrumbsService } from './components/breadcrumbs/service/terra-breadcrumbs.service';
 import { TerraTwoColumnsContainerDirective } from './components/layouts/column-container/terra-two-columns-container.directive';
 
 
@@ -217,7 +215,6 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraDynamicSwitchComponent,
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
-        TerraBreadcrumbsComponent,
         TerraTwoColumnsContainerDirective,
 
         // ### declarations of terra-components examples
@@ -375,7 +372,6 @@ function createCompiler(compilerFactory:CompilerFactory):Compiler
         TerraDynamicSwitchComponent,
         TerraCategoryPickerComponent,
         TerraTwoColumnsContainerComponent,
-        TerraBreadcrumbsComponent,
         TerraTwoColumnsContainerDirective,
 
         // ### declarations of terra-components examples